Add files via upload
The file "Visualization of the traffic lights" is used to visualize the LISA data set to check the correctness. The file "Classification of Traffic lights using a pre-trained TensorFlow model" is used to detect traffic lights using a pre-trained CNN model and then classify the traffic lights into Red and Green by converting the images into HSV domain.